requirements.txt
absl-py==1.4.0
alembic==1.11.1
antlr4-python3-runtime==4.9.3
anyio==3.7.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==23.1.0
autopage==0.5.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
brax==0.9.1
cached-property==1.5.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
chex==0.1.7
click==8.1.3
cliff==4.3.0
cloudpickle==2.2.1
cmaes==0.9.1
cmake==3.28.1
cmd2==2.4.3
colorlog==6.7.0
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dm-env==1.6
dm-tree==0.1.8
docker-pycreds==0.4.0
equinox==0.10.4
etils==1.3.0
evosax==0.1.5
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.12.2
Flask==2.2.5
Flask-Cors==4.0.0
flax @ git+https://github.com/google/flax.git@8995420543bbe9415df5eb7b6a40ed7d87f6db6f
fonttools==4.40.0
fqdn==1.5.1
gitdb==4.0.11
GitPython==3.1.41
glfw==2.6.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
graphviz==0.20.1
greenlet==3.0.3
grpcio==1.56.0
gym==0.26.2
gym-notices==0.0.8
gymnax==0.0.6
h5py==3.9.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
hydra-zen==0.10.1
idna==3.4
imageio==2.31.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.23.3
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.21
jaxlib==0.4.21+cuda12.cudnn89
jaxopt==0.7
jaxtyping==0.2.19
jedi==0.18.2
Jinja2==3.0.3
joblib==1.2.0
jsonpointer==2.4
jsonschema==4.17.3
jumanji==0.3.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
lazy_loader==0.2
lit==17.0.6
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.5
mujoco==2.3.6
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.3.4.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.3.101
nvidia-cuda-nvcc-cu12==12.3.107
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.3.107
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.3.101
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.7.29
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.12.1
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.5.4.101
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.2.0.103
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
objprint==0.2.3
omegaconf==2.3.0
opt-einsum==3.3.0
optax==0.1.4
optuna==2.10.1
orbax-checkpoint==0.2.3
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.8.0
pluggy==1.2.0
prettytable==3.8.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
PyOpenGL==3.1.7
pyparsing==3.1.0
pyperclip==1.8.2
pyrootutils==1.0.4
pyrsistent==0.19.3
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytinyrenderer==0.0.14
pytorch-ignite==0.4.12
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
qdax==0.2.3
qtconsole==5.4.3
QtPy==2.3.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
sentry-sdk==1.40.4
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
stack-data==0.6.2
stevedore==5.1.0
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.1
tensorstore==0.1.39
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trimesh==3.9.35
triton==2.0.0
typeguard==4.0.0
typing_extensions==4.6.3
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
viztracer==0.16.1
wandb==0.16.3
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
widgetsnbextension==4.0.7
zipp==3.15.0
